mongodb_nosql_de/requirements.txt
# Data Pipeline Core
pymongo==4.6.0
pandas==2.1.3
numpy==1.26.2

# Data Quality & Validation
pydantic==2.5.2
jsonschema==4.20.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
requests==2.31.0
boto3==1.34.0

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
mongomock==4.1.2

# Logging & Monitoring
loguru==0.7.2

# API
flask==3.0.0
flask-cors==4.0.0

# Development
black==23.12.1
flake8==6.1.0
mypy==1.7.1
